Fixed a couple of compiler warnings because of misalligned int <-> unint??_t declararions.
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@1620 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/chan_sccp.h
@@ -469,7 +469,7 @@ struct sccp_linedevices {
  */
 
 struct sccp_buttonconfig {
-	uint32_t				instance;				/*!< Instance on device */
+	uint16_t				instance;				/*!< Instance on device */
 	uint16_t				index;					/*!< buttonconfig index */
 	button_type_t 				type;					/*!< Button type (e.g. line, speeddial, feature, empty) */
 	SCCP_LIST_ENTRY(sccp_buttonconfig_t) 	list;					/*!< Button Linked List Entry */
@@ -627,7 +627,7 @@ struct sccp_line {
  */
 struct sccp_speed {
 	uint8_t 				config_instance;			/*!< The instance of the speeddial in the sccp.conf */
-	uint8_t 				instance;				/*!< The instance on the current device */
+	uint16_t 				instance;				/*!< The instance on the current device */
 	uint8_t 				type;					/*!< SpeedDial Button Type (SKINNY_BUTTONTYPE_SPEEDDIAL or SKINNY_BUTTONTYPE_LINE (hint)) */
 	char 					name[StationMaxNameSize];		/*!< The name of the speed dial button */
 	char 					ext[AST_MAX_EXTENSION];			/*!< The number to dial when it's hit */
